requirements.txt
streamlit
google-generativeai
trafilatura